Refactored server start out of the initialization.
